+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add hashtag typeahead completion
 - Cmd-N to create a new note
 - Keyboard support & state restoration for the hashtags
+- Cmd-Enter to enter/exit "edit mode" for the current note
 
 ## [0.38.0] - 2020-01-05